Use session timeout to clear old sandbox
The old strategy allows sessions to remain with a broken sandbox,
because the ns has been removed. This means that after 10 minutes
a user will no longer be able to def new vars and use them. No matter
how long the user waits, or whether they leave the site, they retain the
same session cookie and can no longer def vars.

There are multiple reports of this problem here:

https://github.com/Raynes/tryclojure/issues/64

This change removes the fixed 10 minute timeout to delete the current
namespace, and instead uses a session store that will delete sessions
after 15 minutes of inactivity. The deletion of the session will cause
the deletion of the sandbox, and therefore the ns and associated vars.

This new behaviour actually lives up to what's described in the 'ABOUT'
link.